Refactoring + mine fix

+++ src/main/java/com/nukateam/nukacraft/common/foundation/blocks/blocks/LandMineBlock.java
@@ -1,6 +1,7 @@
 package com.nukateam.nukacraft.common.foundation.blocks.blocks;
 
 import com.nukateam.guns.common.data.util.VoxelShapeHelper;
+import com.nukateam.nukacraft.common.foundation.blocks.entity.OwnableBlockEntity;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
@@ -9,17 +10,22 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Explosion;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.BaseEntityBlock;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.EntityBlock;
+import net.minecraft.world.level.block.RenderShape;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
+import net.minecraftforge.common.MinecraftForge;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
 
-public class LandMineBlock extends Block {
+public class LandMineBlock extends BaseEntityBlock {
     private final Map<BlockState, VoxelShape> SHAPES = new HashMap<>();
-    private UUID owner;
+//    private UUID owner;
 
     public LandMineBlock(Properties pProperties) {
         super(pProperties);
@@ -27,12 +33,21 @@ public class LandMineBlock extends Block {
 
     @Override
     public void entityInside(BlockState pState, Level pLevel, BlockPos pPos, Entity pEntity) {
-        if (!pLevel.isClientSide) {
-            if (!(owner == pEntity.getUUID())) {
+
+        var blockEntity = pLevel.getBlockEntity(pPos);
+        if (!pLevel.isClientSide && blockEntity instanceof OwnableBlockEntity ownable) {
+            if (!(ownable.getOwner().equals(pEntity.getUUID().toString()))) {
                 pLevel.destroyBlock(pPos, false);
                 pLevel.explode(null, pPos.getX(), pPos.getY(),pPos.getZ(),6.0f, Explosion.BlockInteraction.NONE);
             }
         }
+
+//        if (!pLevel.isClientSide) {
+//            if (!(owner == pEntity.getUUID())) {
+//                pLevel.destroyBlock(pPos, false);
+//                pLevel.explode(null, pPos.getX(), pPos.getY(),pPos.getZ(),6.0f, Explosion.BlockInteraction.NONE);
+//            }
+//        }
     }
 
     @Override
@@ -49,9 +64,24 @@ public class LandMineBlock extends Block {
 
     @Override
     public void setPlacedBy(Level pLevel, BlockPos pPos, BlockState pState, @Nullable LivingEntity pPlacer, ItemStack pStack) {
-        if (pPlacer instanceof Player) {
-            owner = pPlacer.getUUID();
+        if (pPlacer instanceof Player player) {
+//            owner = pPlacer.getUUID();
+            var uuid = player.getGameProfile().getId().toString();
+            var entity = pLevel.getBlockEntity(pPos);
+            if (entity instanceof OwnableBlockEntity ownable) {
+                ownable.setOwner(uuid);
+            }
         }
         super.setPlacedBy(pLevel, pPos, pState, pPlacer, pStack);
+    }
+
+    @Override
+    public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
+        return new OwnableBlockEntity(pos, state);
+    }
+
+    @Override
+    public RenderShape getRenderShape(BlockState pState) {
+        return RenderShape.MODEL;
     }
 }
